@@ -1,32 +1,27 @@
-{
-  "name": "travel-diary-app",
-  "version": "0.1.0",
+﻿{
+  "name": "travel-diary-spa",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "migrate": "prisma migrate dev --name init",
-    "generate": "prisma generate",
-    "seed": "ts-node prisma/seed.ts"
+    "dev": "vite dev",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@prisma/client": "5.0.0",
-    "next": "14.1.0",
-    "prisma": "5.0.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "@supabase/supabase-js": "^2.42.0",
+    "exifr": "^7.1.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.26.1"
   },
   "devDependencies": {
-    "@types/node": "20.4.2",
-    "@types/react": "18.2.14",
-    "autoprefixer": "10.4.15",
-    "eslint": "8.45.0",
-    "eslint-config-next": "14.1.0",
-    "postcss": "8.4.31",
-    "tailwindcss": "3.4.0",
-    "ts-node": "10.9.1",
-    "typescript": "5.2.2"
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.41",
+    "tailwindcss": "^3.4.10",
+    "typescript": "^5.5.4",
+    "vite": "^5.4.2"
   }
 }
